Updated compile-line arguments in CMakeLists.txt to reduce the number of warnings printed under MSVC

+++ src/ZGPeerSession.cpp
@@ -59,7 +59,7 @@ static ZGPeerID GenerateLocalPeerID()
 #endif
 
    // And finally some additional salt just to make use of the other 32 low-bits.  This probably isn't necessary, but I'm paranoid
-   unsigned int seed = time(NULL);
+   unsigned int seed = (unsigned int) time(NULL);
    const uint32 salt = CalculateHashCode(GetCurrentTime64() + GetRunTime64() + GetRandomNumber(&seed));
    return ZGPeerID((macAddress<<16)|((uint64)GetNextUniqueObjectID()), (((uint64)processID)<<32)|((uint64)salt));
 }
